pyproject.toml
[project]
name = "vectorless-rag"
version = "0.1.0"
description = "Vectorless RAG system using FastAPI and modern AI tools"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.119.1",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "jsonpatch>=1.33",
    "jsonschema>=4.25.1",
    "minio>=7.2.18",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.0.0",
    "pydantic-ai>=1.3.0",
    "pymongo>=4.15.3",
    "pymupdf>=1.26.5",
    "pypdf>=6.1.3",
    "pypdf2>=3.0.1",
    "pytest>=8.4.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=7.0.0",
    "requests>=2.32.5",
    "structlog>=25.4.0",
    "uvicorn>=0.38.0",
]
